AI generation/requirements.txt
pyyaml
markdown-it-py

